pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "anemll"
version = "0.1.0"
description = "Run LLMs on Apple Neural Engine with an OpenAI-compatible API"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "pydantic>=2.10.0",
    "torch>=2.6.0",
    "numpy>=2.0.0",
    "transformers>=4.49.0",
    "coremltools>=8.2",
    "PyYAML>=6.0.0",
    "python-multipart>=0.0.20",
]

[project.scripts]
anemll = "anemll.cli:main"

[tool.setuptools.packages.find]
include = ["anemll*"]
